Demo Whitebox Ready
Fixed cursor issue, now works with gamepad or mouse and keyboard. Cleaned up some spawning bugs as well as tweaked some upgrade values for more engaging gameplay.
Next steps will be to included more player stats and effects as well as build out a few more test levels for more variation when testing.
Also should polish up SpawnCreator so that it stores the point of contact, the normal from the surface for each point. As well store the maps transform so that the points will auto update if the map gets altered. Should also then include the map in the spawnPattern data...